fix breaking rename class

+++ src/WinTenBot/WinTenBot/Helpers/BotHelper.cs
@@ -10,6 +10,7 @@ using Telegram.Bot.Types;
 using Telegram.Bot.Types.InputFiles;
 using WinTenBot.Model;
 using WinTenBot.Providers;
+using WinTenBot.Services;
 using File = System.IO.File;
 
 namespace WinTenBot.Helpers
@@ -30,13 +31,13 @@ namespace WinTenBot.Helpers
 
         public static async Task<bool> IsBotAdded(this User[] users)
         {
-            var me = await Bot.Client.GetMeAsync();
+            var me = await BotSettings.Client.GetMeAsync();
             return (from user in users where user.Id == me.Id select user.Id == me.Id).FirstOrDefault();
         }
 
         public static bool IsRestricted()
         {
-            return Bot.GlobalConfiguration["CommonConfig:IsRestricted"].ToBool();
+            return BotSettings.GlobalConfiguration["CommonConfig:IsRestricted"].ToBool();
         }
 
         public static bool CheckRestriction(this long chatId)
@@ -45,7 +46,7 @@ namespace WinTenBot.Helpers
             {
                 var isRestricted = false;
                 var globalRestrict = IsRestricted();
-                var sudoers = Bot.GlobalConfiguration.GetSection("RestrictArea").Get<List<string>>();
+                var sudoers = BotSettings.GlobalConfiguration.GetSection("RestrictArea").Get<List<string>>();
                 var match = sudoers.FirstOrDefault(x => x == chatId.ToString());
 
                 Log.Information($@"Global Restriction: {globalRestrict}");
@@ -83,39 +84,69 @@ namespace WinTenBot.Helpers
 
         public static ITelegramBotClient GetClient(string name)
         {
-            return Bot.Clients[name];
+            return BotSettings.Clients[name];
         }
 
         public static async Task ClearLog()
         {
-            const string logsPath = "Storage/Logs";
-            var botClient = Bot.Client;
-            var channelTarget = Bot.GlobalConfiguration["CommonConfig:ChannelLogs"].ToInt64();
-
-            var dirInfo = new DirectoryInfo(logsPath);
-            var files = dirInfo.GetFiles();
-            var filteredFiles = files.Where(fileInfo =>
-                fileInfo.CreationTimeUtc < DateTime.UtcNow.AddDays(-1)).ToArray();
-
-            if (filteredFiles.Length > 0)
+            try
             {
-                Log.Information($"Found {filteredFiles.Length} of {files.Length}");
-                foreach (var fileInfo in filteredFiles)
+                const string logsPath = "Storage/Logs";
+                var botClient = BotSettings.Client;
+                var channelTarget = BotSettings.GlobalConfiguration["CommonConfig:ChannelLogs"].ToInt64();
+
+                if (!channelTarget.ToString().StartsWith("-100"))
                 {
-                    var filePath = fileInfo.FullName;
-                    Log.Information($"Uploading file {filePath}");
-                    var fileStream = File.OpenRead(filePath);
-
-                    var media = new InputOnlineFile(fileStream, fileInfo.Name);
-                    await botClient.SendDocumentAsync(channelTarget, media);
-
-                    filePath.DeleteFile();
+                    Log.Information("Please specify ChannelTarget in appsettings.json");
+                    return;
                 }
+
+                var dirInfo = new DirectoryInfo(logsPath);
+                var files = dirInfo.GetFiles();
+                var filteredFiles = files.Where(fileInfo =>
+                    fileInfo.CreationTimeUtc < DateTime.UtcNow.AddDays(-1)).ToArray();
+
+                if (filteredFiles.Length > 0)
+                {
+                    Log.Information($"Found {filteredFiles.Length} of {files.Length}");
+                    foreach (var fileInfo in filteredFiles)
+                    {
+                        var filePath = fileInfo.FullName;
+                        Log.Information($"Uploading file {filePath}");
+                        var fileStream = File.OpenRead(filePath);
+
+                        var media = new InputOnlineFile(fileStream, fileInfo.Name);
+                        await botClient.SendDocumentAsync(channelTarget, media);
+
+                        filePath.DeleteFile();
+                    }
+                }
+                else
+                {
+                    Log.Information("No Logs file need be processed for previous date");
+                }
+            }catch(Exception ex){
+                Log.Error(ex, "Error Send .Log file to ChannelTarget.");
             }
-            else
+        }
+
+        public static async Task EnsureChatHealthAsync(this TelegramProvider telegramProvider)
+        {
+            Log.Information("Ensuring chat health..");
+            
+            var message = telegramProvider.Message;
+            var settingsService = new SettingsService
             {
-                Log.Information("No Logs file need be processed for previous date");
-            }
+                Message = message
+            };
+
+            var data = new Dictionary<string, object>()
+            {
+                ["chat_id"] = message.Chat.Id,
+                ["chat_title"] = message.Chat.Title,
+            };
+                
+            var update = await settingsService.SaveSettingsAsync(data);
         }
     }
 }
